[feat] #123 ChatDetailViewModel 생성
